clean some classes

+++ laravel/app/Http/Controllers/DepartmentController.php
@@ -9,6 +9,7 @@ use App\Http\Resources\DepartmentResource;
 use App\Models\Department;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Auth;
+use App\Helpers\UserRole;
 class DepartmentController extends Controller
 {
 
@@ -20,24 +21,34 @@ class DepartmentController extends Controller
 
     public function store(DepartmentRequest $request):JsonResponse
     {
-        $department= Department::create(array_merge($request->all(), ['admin_id' => Auth::id()]));
-        return $this->success(new DepartmentResource($department) , 'We save the Department');
+        if (UserRole::isAdmin()) {
+            $department = Department::create(array_merge($request->all(), ['admin_id' => Auth::id()]));
+            return $this->success(new DepartmentResource($department), 'We save the Department');
+        }
+        return $this->error(null, 'You do not have the Auth to create a department');
     }
 
     public function show(Department $department):JsonResponse
     {
-        return $this->success(new DepartmentResource($department) , 'This is specific Department');
+        return $this->success(new DepartmentResource($department) , 'This is specific ' .$department->name);
     }
 
     public function update(UpdateDepartment $request, Department $department):JsonResponse
     {
-        $department->update(array_merge($request->all(), ['admin_id' => Auth::id()]));
-        return $this->success(new DepartmentResource($department) , 'We update the department');
+        if (UserRole::isAdmin()) {
+            $department->update(array_merge($request->all(), ['admin_id' => Auth::id()]));
+            return $this->success(new DepartmentResource($department), 'We update the '.$department->name);
+        }
+        return $this->error(null, 'You do not have the Auth to update the department');
     }
 
     public function destroy(Department $department):JsonResponse
     {
-        $department->delete();
-        return $this->success(null , 'We delete the department');
+        if(UserRole::isAdmin()) {
+            $department->delete();
+            return $this->success(null, 'We delete the department');
+        }
+        return $this->error(null, 'You do not have the Auth to delete the department');
+
     }
 }
